+++ flow_test.go
@@ -70,16 +70,16 @@ func Test_successful(t *testing.T) {
 		return []int{executed1, executed2, executed3}
 	}
 
-	exitCode := flow.Execute(ctx, "task-1")
-	requireEqual(t, exitCode, 0, "first execution should pass")
+	err := flow.Execute(ctx, "task-1")
+	requireEqual(t, err, nil, "first execution should pass")
 	requireEqual(t, got(), []int{1, 0, 0}, "should execute task 1")
 
-	exitCode = flow.Execute(ctx, "task-2")
-	requireEqual(t, exitCode, 0, "second execution should pass")
+	err = flow.Execute(ctx, "task-2")
+	requireEqual(t, err, nil, "second execution should pass")
 	requireEqual(t, got(), []int{2, 1, 0}, "should execute task 1 and 2")
 
-	exitCode = flow.Execute(ctx, "task-1", "task-2", "task-3")
-	requireEqual(t, exitCode, 0, "third execution should pass")
+	err = flow.Execute(ctx, "task-1", "task-2", "task-3")
+	requireEqual(t, err, nil, "third execution should pass")
 	requireEqual(t, got(), []int{3, 2, 1}, "should execute task 1 and 2 and 3")
 }
 
@@ -116,9 +116,9 @@ func Test_dependency_failure(t *testing.T) {
 		return []int{executed1, executed2, executed3}
 	}
 
-	exitCode := flow.Execute(context.Background(), "task-2", "task-3")
+	err := flow.Execute(context.Background(), "task-2", "task-3")
 
-	assertEqual(t, exitCode, 1, "should return error from first task")
+	assertFail(t, err, "should return error from first task")
 	assertEqual(t, got(), []int{11, 0, 0}, "should execute task 1")
 }
 
@@ -135,9 +135,9 @@ func Test_fail(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(context.Background(), "task")
+	err := flow.Execute(context.Background(), "task")
 
-	assertEqual(t, exitCode, 1, "should return error")
+	assertFail(t, err, "should return error")
 	assertTrue(t, failed, "tf.Failed() should return true")
 }
 
@@ -154,9 +154,9 @@ func Test_skip(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(context.Background(), "task")
+	err := flow.Execute(context.Background(), "task")
 
-	assertEqual(t, exitCode, 0, "should pass")
+	assertPass(t, err, "should pass")
 	assertTrue(t, skipped, "tf.Skipped() should return true")
 }
 
@@ -186,9 +186,9 @@ func Test_task_panics(t *testing.T) {
 				Action: tc.action,
 			})
 
-			exitCode := flow.Execute(context.Background(), "task")
+			err := flow.Execute(context.Background(), "task")
 
-			assertEqual(t, exitCode, 1, "should return error from first task")
+			assertFail(t, err, "should return error from first task")
 		})
 	}
 }
@@ -201,9 +201,9 @@ func Test_cancelation(t *testing.T) {
 		Name: "task",
 	})
 
-	exitCode := flow.Execute(ctx, "task")
+	err := flow.Execute(ctx, "task")
 
-	assertEqual(t, exitCode, 1, "should return error canceled")
+	assertEqual(t, err, context.Canceled, "should return error canceled")
 }
 
 func Test_cancelation_during_last_task(t *testing.T) {
@@ -217,9 +217,9 @@ func Test_cancelation_during_last_task(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(ctx, "task")
+	err := flow.Execute(ctx, "task")
 
-	assertEqual(t, exitCode, 0, "should pass as the flow completed")
+	assertPass(t, err, "should pass as the flow completed")
 }
 
 func Test_empty_action(t *testing.T) {
@@ -228,9 +228,9 @@ func Test_empty_action(t *testing.T) {
 		Name: "task",
 	})
 
-	exitCode := flow.Execute(context.Background(), "task")
+	err := flow.Execute(context.Background(), "task")
 
-	assertEqual(t, exitCode, 0, "should pass")
+	assertPass(t, err, "should pass")
 }
 
 func Test_invalid_args(t *testing.T) {
@@ -253,9 +253,9 @@ func Test_invalid_args(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.desc, func(t *testing.T) {
-			exitCode := flow.Execute(context.Background(), tc.args...)
+			err := flow.Execute(context.Background(), tc.args...)
 
-			assertEqual(t, exitCode, 2, "should return error bad args")
+			assertInvalid(t, err, "should return error bad args")
 		})
 	}
 }
@@ -283,7 +283,7 @@ func Test_printing(t *testing.T) {
 		},
 	})
 
-	flow.Execute(context.Background(), "failing")
+	_ = flow.Execute(context.Background(), "failing")
 
 	assertContains(t, out, "Skipf 0", "should contain proper output from \"skipped\" task")
 	assertContains(t, out, "Fatalf 5", "should contain proper output from \"failing\" task")
@@ -307,9 +307,9 @@ func Test_concurrent_printing(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(context.Background(), "task")
+	err := flow.Execute(context.Background(), "task")
 
-	assertEqual(t, exitCode, goyek.CodeFail, "should fail")
+	assertFail(t, err, "should fail")
 	assertContains(t, out, "from child goroutine", "should contain log from child goroutine")
 	assertContains(t, out, "from main goroutine", "should contain log from main goroutine")
 }
@@ -342,9 +342,9 @@ func Test_concurrent_error(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(context.Background(), "task")
+	err := flow.Execute(context.Background(), "task")
 
-	assertEqual(t, exitCode, goyek.CodeFail, "should fail")
+	assertFail(t, err, "should fail")
 }
 
 func Test_name(t *testing.T) {
@@ -358,9 +358,8 @@ func Test_name(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(context.Background(), taskName)
+	_ = flow.Execute(context.Background(), taskName)
 
-	assertEqual(t, exitCode, 0, "should pass")
 	assertEqual(t, got, taskName, "should return proper Name value")
 }
 
@@ -375,7 +374,7 @@ func Test_output(t *testing.T) {
 		},
 	})
 
-	flow.Execute(context.Background(), "task")
+	_ = flow.Execute(context.Background(), "task")
 
 	assertContains(t, out, msg, "should contain message send via output")
 }
@@ -391,9 +390,9 @@ func Test_SetDefault(t *testing.T) {
 	})
 	flow.SetDefault(task)
 
-	exitCode := flow.Execute(context.Background())
+	err := flow.Execute(context.Background())
 
-	assertEqual(t, exitCode, 0, "should pass")
+	assertPass(t, err, "should pass")
 	assertTrue(t, taskRan, "task should have run")
 }
 
@@ -424,10 +423,10 @@ func TestCmd_success(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(context.Background(), taskName)
+	err := flow.Execute(context.Background(), taskName)
 
+	assertPass(t, err, "task should pass")
 	assertContains(t, out, "go version go", "output should contain prefix of version report")
-	assertEqual(t, exitCode, goyek.CodePass, "task should pass")
 }
 
 func TestCmd_error(t *testing.T) {
@@ -442,9 +441,9 @@ func TestCmd_error(t *testing.T) {
 		},
 	})
 
-	exitCode := flow.Execute(nil, taskName) //nolint:staticcheck // present that nil context is handled
+	err := flow.Execute(nil, taskName) //nolint:staticcheck // present that nil context is handled
 
-	assertEqual(t, exitCode, goyek.CodeFail, "task should pass")
+	assertFail(t, err, "task should pass")
 }
 
 func TestFlow_Tasks(t *testing.T) {
@@ -496,28 +495,6 @@ func TestFlow_Print(t *testing.T) {
 	assertContains(t, out, "(depends on: task)", "should print the task dependencies")
 }
 
-func TestFlow_Usage_default(t *testing.T) {
-	out := &strings.Builder{}
-	flow := &goyek.Flow{Output: out}
-	flow.Define(goyek.Task{Name: "task"})
-
-	exitCode := flow.Execute(context.Background(), "bad")
-
-	assertEqual(t, exitCode, goyek.CodeInvalidArgs, "should work when invalid code returned")
-	assertContains(t, out, "Tasks:", "should print the default usage if not overridden")
-}
-
-func TestFlow_Usage_custom(t *testing.T) {
-	flow := &goyek.Flow{Output: &strings.Builder{}}
-	called := false
-	flow.Usage = func() { called = true }
-
-	exitCode := flow.Execute(context.Background(), "bad")
-
-	assertEqual(t, exitCode, goyek.CodeInvalidArgs, "should work when invalid code returned")
-	assertTrue(t, called, "should invoke the custom message instead")
-}
-
 func TestFlow_Logger(t *testing.T) {
 	out := &strings.Builder{}
 	flow := &goyek.Flow{Output: out, Logger: goyek.FmtLogger{}}
@@ -529,7 +506,7 @@ func TestFlow_Logger(t *testing.T) {
 		},
 	})
 
-	flow.Execute(context.Background(), "task")
+	_ = flow.Execute(context.Background(), "task")
 
 	assertContains(t, out, "first", "should call Log")
 	assertContains(t, out, "second", "should call Logf")
@@ -548,7 +525,7 @@ func TestFlow_Use(t *testing.T) {
 		}
 	})
 
-	flow.Execute(context.Background(), "task")
+	_ = flow.Execute(context.Background(), "task")
 
 	assertContains(t, out, "message", "should call middleware with proper input")
 }
@@ -603,6 +580,27 @@ func assertEqual(tb testing.TB, got interface{}, want interface{}, msg string) {
 		return
 	}
 	tb.Errorf("%s\nGOT: %v\nWANT: %v", msg, got, want)
+}
+
+func assertPass(tb testing.TB, got error, msg string) {
+	tb.Helper()
+	if got != nil {
+		tb.Errorf("%s\nGOT: %v\nWANT: <PASS>", msg, got)
+	}
+}
+
+func assertFail(tb testing.TB, got error, msg string) {
+	tb.Helper()
+	if _, ok := got.(*goyek.FailError); !ok {
+		tb.Errorf("%s\nGOT: %v\nWANT: <FAIL>", msg, got)
+	}
+}
+
+func assertInvalid(tb testing.TB, got error, msg string) {
+	tb.Helper()
+	if _, ok := got.(*goyek.FailError); ok || got == nil {
+		tb.Errorf("%s\nGOT: %v\nWANT: <INVALID>", msg, got)
+	}
 }
 
 func assertPanics(tb testing.TB, fn func(), msg string) {